% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_valid.R
\name{impute_valid}
\alias{impute_valid}
\title{multivariable Time Series Data performance evaluation method}
\usage{
impute_valid(data, imputeData, NAdata, col_name = NULL)
}
\arguments{
\item{NA}{,b : two numbers to be operated}
}
\description{
This function is a package created based on the function of 'spTimer' packages algorithm
}
\examples{
# imputation value return
impute=imputeTSRF(NAdata,return_kalman = T)
# Performance evaluation is performed for cases where NAdata is missing for each column.
idx1=impute_valid(data,impute[[1]],NAdata)
}
\keyword{a}
\keyword{b}
\keyword{from}
\keyword{subtract}
