% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeTSRF.R
\name{imputeTSRF}
\alias{imputeTSRF}
\title{multivariable Time Series Data Imputation method}
\usage{
imputeTSRF(
  NAdata,
  return_kalman = F,
  return_model = T,
  smooth = T,
  seed = 1,
  ...
)
}
\arguments{
\item{'NAdata'}{is data with missing columns in multiple columns}
}
\description{
This function is a package created based on the function of 'imputeTS', 'MissForest' packages algorithm
}
\examples{
# imputation value return
impute=imputeTSRF(NAdata,return_kalman = T)
# Performance evaluation is performed for cases where NAdata is missing for each column.
idx1=impute_valid(data,impute[[1]],NAdata)
}
